--created multiple files for the code-- 1. GYRO.cpp added. it contains all of the gyro functions 2. DECLARATIONS.h added. it contains all of the necessary function declarations and some of the include files.
*note: the arduino ide is honky. see: http://jamesreubenknowles.com/including-libraries-in-the-arduino-ide-1625
Any Arduino libraries that are needed by the files outside of the sketch (.ino) file must also be listed in the sketch file itself.
The sketch is parsed for include files. The sketch, all included header files, and the corresponding source files, are copied to another directory for compiling. From that directory, library-based include files are NOT available unless they are included in the sketch and copied to the build directory.

**note: this code compiles, but has not been road/functionally tested.