***fail_safe.ino is fixed*** glitchy servo and non-working ESC are now working because of the following...
the fail_safe.ino wasn't working correctly because it would always switch to MANUAL whenever a pulse was received from the RX. That means, every 20ms, the muxer would switch to manual, set the pins, and the switch back to the appropriate state.

now, the fail_safe.ino does the following... it checks to see if the mode has changed. if it has NOT, it skips the loop. if it HAS, the muxer is set to manual, the pins are changed, and the muxer is set back to the appropriate state.

Signed-off-by: rkburnside <richard@burnsides.org>